refactor: rename folder name and delete redundant code
