+++ skirunmap/src/auth/Firebase/index.ts
@@ -2,7 +2,6 @@ import { getFirestore } from '@firebase/firestore'
 import { initializeApp } from 'firebase/app'
 import { getAuth } from 'firebase/auth'
 import { getStorage } from 'firebase/storage'
-// import { getAnalytics } from "firebase/analytics";
 
 const firebaseConfig = {
   apiKey: 'AIzaSyC_F8P7qOtAZ-go0RGE9OIoqXRB4VjHMr8',
@@ -15,15 +14,8 @@ const firebaseConfig = {
 }
 
 const app = initializeApp(firebaseConfig)
-// const analytics = getAnalytics(app);
-
-/* Firestore */
 const db = getFirestore(app)
-
-/* Storage */
 const storage = getStorage(app)
-
-/* Authentication */
 const auth = getAuth(app)
 
 export { app, auth, db, storage }
