+++ skirunmap/src/pages/Home/index.tsx
@@ -8,7 +8,7 @@ import {
   where
 } from 'firebase/firestore'
 import { useEffect, useState } from 'react'
-import { db } from '../../auth/CloudStorage'
+import { db } from '../../auth/Firebase'
 import FilterSVG from '../../components/FilterSVG'
 import HeroHeader from '../../components/HeroHeader'
 import RouteCard from '../../components/RouteCard'
