+++ skirunmap/src/pages/RouteEdit/index.tsx
@@ -13,7 +13,7 @@ import { useBlocker, useNavigate } from 'react-router-dom'
 import { toast } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css'
 import { v4 as uuidv4 } from 'uuid'
-import { db, storage } from '../../auth/CloudStorage'
+import { db, storage } from '../../auth/Firebase'
 import Map from '../../components/Map'
 import Modal from '../../components/Modal'
 import { MarkerWithSpotId, useMapStore } from '../../store/useMap'
